pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = "User service for a microservices architecture."
authors = ["Amir Sadati eng.amirh.sadati@gmail.com"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
pydantic = {extras = ["email"], version = "^2.0.0"}
pydantic-settings = "^2.0.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.19"}
asyncpg = "^0.28.0"
alembic = "^1.11.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6" # For form data
gunicorn = "^21.2.0"
prometheus-fastapi-instrumentator = "^6.1.0"
opentelemetry-sdk = "^1.19.0"
opentelemetry-instrumentation-fastapi = "^0.40b0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.24.1" # For async test client

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
